rework session view; separate toolbar dir
