+++ src/plugins/spark/components/GenericBlock/UnknownBlockWidget.vue
@@ -54,7 +54,7 @@ export default class UnknownBlockWidget extends WidgetBase {
 
 <template>
   <q-card dark class="text-white scroll">
-    <WidgetToolbar :title="widget.title" :subtitle="displayName">
+    <WidgetToolbar :crud="crud">
       <q-item-section side>
         <q-btn-dropdown flat label="Menu">
           <q-list dark bordered>
