+++ src/plugins/spark/features/SessionView/index.ts
@@ -1,7 +1,6 @@
-import { ref } from '@/helpers/component-ref';
+import { ref, selector } from '@/helpers/component-ref';
 import { Feature } from '@/store/features';
 
-import form from './SessionViewForm.vue';
 import widget from './SessionViewWidget.vue';
 import wizard from './SessionViewWizard.vue';
 import { SessionViewConfig } from './types';
@@ -9,7 +8,7 @@ import { SessionViewConfig } from './types';
 const feature: Feature = {
   id: 'SessionView',
   displayName: 'Session View',
-  widgetComponent: ref(widget),
+  widgetComponent: selector(widget),
   wizardComponent: ref(wizard),
   validator: (config: SessionViewConfig) => !!config.sessions,
   widgetSize: {
